requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests-oauthlib==1.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
